Commited.
Online Bus Reservation System is the project whose objective is to make the seat reservation online so that end-user can book tickets from their own place. I have made this project in Java GUI. I am providing you screenshots with descriptions for this Java Project. If you want to see the demo of this project you can see this at the bottom of this post. If you would like to check more computer science projects.